:bricks: move css and fonts (again) into assets
