+++ src/assets/TODO colocate image files next to their use
